P4
On the event planning app, I was assigned the task to enable the view details link for each event. I created a page, added some CSS styling to it, and activated the link to redirect to that page once clicked. I worked together with my team to troubleshoot issues with other colleague features - we worked extensively with creating files needed for the search feature.
Overall, we are moving on to new tasks of adding additional features such as separating user roles into volunteer and organizational user. We intend to give them different authorities. I will be working on the view seen when user is in session.